remove problemtic ignore_windows support from lnotify plugin
The python "lnotify" plugin had some problematic code that would
suppress notifications if the active window was named "tilda",
"xterm", or "gnome-terminal". This list was not user-configurable, and
the code implementing this feature was broken due to #339
(incompatible with common versions of ps) and #334 (python3
bytes-vs-string incompatibility).

This commit removes the problematic code.

Closes: #334
Closes: #339